refactor(swapper): extract shared starknet-utils
Move the multi-call INVOKE building flow (toHexString, RPC fee estimate,
resourceBounds calculation, invoke-hash construction) into a shared
packages/swapper/src/starknet-utils/ module. AvnuSwapper and GardenSwapper
now both call buildStarknetInvokeTx with their pre-formatted calldata,
keeping each swapper responsible only for assembling its own Call array.

No behaviour change for Avnu — the extracted code is byte-equivalent to
its previous inline implementation. Garden's endpoints.ts shrinks
from 333 to ~140 lines and the duplication that was flagged in code
review is eliminated.
